update endpoint tests
